requirements.txt
build==1.2.1
pytest==8.3.2
pyzmq==26.0.3
setuptools==72.0.0
twine==5.1.1
uv==0.2.30
protobuf==5.27.3
betterproto==1.2.5
